pyproject.toml
[project]
name = "anthropic-mastery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "python-dotenv>=1.1.1",
]
